CADC-11734 Moved /stats to /session?view=stats
